requirements.txt
# Core dependencies
fastapi==0.104.0
uvicorn==0.23.2
pydantic==1.10.8
email-validator==2.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.22
asyncpg==0.28.0
alembic==1.12.0

# Authentication
python-jose==3.3.0
passlib==1.7.4

# Cloud storage
google-cloud-storage==2.10.0
boto3==1.28.38

# Image processing
pillow==10.0.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
httpx==0.25.0

# Visualization
matplotlib==3.7.2
plotly==5.18.0
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0